+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 [![Donate via PayPal](https://img.shields.io/badge/Donate-PayPal-blue.svg?style=flat-square)](https://www.paypal.me/techtoday) 
 
 <p>
+<b>Version 1.1.4</b><br/>
+- Last changed status date/time shortened out<br/>
+</p>
+<p>
 <b>Version 1.1.3</b><br/>
 - In node status, added the last changed status date/time.<br/>
 </p>